refactor amp/bento-accordion
